+++ language-server/src/tests/action.rs
@@ -15964,3 +15964,146 @@ fn convert_to_int_has_nicely_separated_digits() {
         find_position_of("0b100101101011010000111").to_selection()
     );
 }
+
+// https://github.com/gleam-lang/gleam/issues/6120
+#[test]
+fn discard_unused_variable_triggered_at_string_prefix_pattern() {
+    assert_code_action!(
+        DISCARD_UNUSED_VARIABLE,
+        r#"fn wobble(wibble) {
+  case wibble {
+    "Hello" as hello <> name -> name
+    _ -> "unknown"
+  }
+}"#,
+        find_position_of("hello").to_selection()
+    );
+}
+
+// https://github.com/gleam-lang/gleam/issues/6121
+#[test]
+fn discard_unused_variable_triggered_at_alternative_pattern_list() {
+    assert_code_action!(
+        DISCARD_UNUSED_VARIABLE,
+        r#"fn wobble(wibble) {
+  case wibble {
+    [0, x, y] | [1, x, y] | [2, x, y] -> x
+    _ -> 0
+  }
+}"#,
+        find_position_of("y").to_selection()
+    );
+}
+
+// https://github.com/gleam-lang/gleam/issues/6121
+#[test]
+fn discard_unused_variable_triggered_at_alternative_pattern_string_prefix() {
+    assert_code_action!(
+        DISCARD_UNUSED_VARIABLE,
+        r#"fn wobble(wibble) {
+  case wibble {
+    "Hello, " as hello <> name | "Hey, " as hello <> name -> hello
+    _ -> "unknown"
+  }
+}"#,
+        find_position_of("name").to_selection()
+    );
+}
+
+// https://github.com/gleam-lang/gleam/issues/6121
+#[test]
+fn discard_unused_variable_triggered_at_alternative_pattern_string_prefix_left_side() {
+    assert_code_action!(
+        DISCARD_UNUSED_VARIABLE,
+        r#"fn wobble(wibble) {
+  case wibble {
+    "Hello, " as hello <> name | "Hey, " as hello <> name -> name
+    _ -> "unknown"
+  }
+}"#,
+        find_position_of("hello").to_selection()
+    );
+}
+
+// https://github.com/gleam-lang/gleam/issues/6121
+#[test]
+fn discard_unused_variable_triggered_at_alternative_pattern_multiple_mixed() {
+    assert_code_action!(
+        DISCARD_UNUSED_VARIABLE,
+        r#"fn wobble(wibble, wubble) {
+  case wibble, wubble {
+    0, [y] | y, [] | _, [1, 2, y] -> todo
+    _ -> todo
+  }
+}"#,
+        find_position_of("y").to_selection()
+    );
+}
+
+// https://github.com/gleam-lang/gleam/issues/6121
+#[test]
+fn discard_unused_variable_triggered_at_alternative_pattern_list_assignments() {
+    assert_code_action!(
+        DISCARD_UNUSED_VARIABLE,
+        r#"fn wobble(wibble) {
+  case wibble {
+    [0, ..y] | [0, 1, ..] as y | y -> todo
+    _ -> todo
+  }
+}"#,
+        find_position_of("y").to_selection()
+    );
+}
+
+// https://github.com/gleam-lang/gleam/issues/6121
+#[test]
+fn discard_unused_variable_triggered_at_list_pattern_assignment_with_alternative_patterns() {
+    assert_code_action!(
+        DISCARD_UNUSED_VARIABLE,
+        "fn wibble(wobble: List(Int)) -> Nil {
+  case wobble {
+    [1, 2, ..] as tail | [2, 1, ..] as tail -> Nil
+    _ -> Nil
+  }
+}",
+        find_position_of("tail").to_selection()
+    );
+}
+
+// https://github.com/gleam-lang/gleam/issues/6121
+#[test]
+fn discard_unused_variable_triggered_at_pattern_assignment_with_alternative_patterns() {
+    assert_code_action!(
+        DISCARD_UNUSED_VARIABLE,
+        "pub type Wibble {
+  Wibble(a: Int)
+  Wobble(a: Int)
+}
+
+fn wobble(wibble: Wibble) -> Int {
+  case wibble {
+    Wibble(a) | Wobble(a) -> 0
+  }
+}",
+        find_position_of("a").nth_occurrence(4).to_selection()
+    );
+}
+
+// https://github.com/gleam-lang/gleam/issues/6121
+#[test]
+fn discard_unused_variable_triggered_at_pattern_assignment_shorthand_with_alternative_patterns() {
+    assert_code_action!(
+        DISCARD_UNUSED_VARIABLE,
+        "pub type Wibble {
+  Wibble(a: Int)
+  Wobble(a: Int)
+}
+
+fn wobble(wibble: Wibble) -> Int {
+  case wibble {
+    Wibble(a:) | Wobble(a:) -> 0
+  }
+}",
+        find_position_of("a:").nth_occurrence(3).to_selection()
+    );
+}